feat(Layout): add story
